crash [@ nsMsgDatabase::DeleteHeader(nsIMsgDBHdr*, nsIDBChangeListener*, int, int)] with pop account, plus some random totally unrelated corrections to non-executable files with executable permissions, r=bienvenu, sr=Standard8
--HG--
extra : rebase_source : 5cdc012ecf5feaafd57200d14666fd57c0afe609

+++ mailnews/local/src/nsPop3Sink.cpp
@@ -248,7 +248,7 @@ nsPop3Sink::CheckPartialMessages(nsIPop3Protocol *protocol)
     PRBool found = PR_TRUE;
     partialMsg = static_cast<partialRecord *>(m_partialMsgsArray.ElementAt(i));
     protocol->CheckMessage(partialMsg->m_uidl.get(), &found);
-    if (!found)
+    if (!found && partialMsg->m_msgDBHdr)
     {
       m_newMailParser->m_mailDB->DeleteHeader(partialMsg->m_msgDBHdr, nsnull, PR_FALSE, PR_TRUE);
       deleted = PR_TRUE;
